style(ruff): B007 + RUF005 + RUF046 — 19 sites, no behaviour change
Three rules that ruff marks unsafe-to-autofix but which are provably
value-identical here, so they were done by hand and checked one by one:

  B007  (5) unused loop targets renamed to `_name`. In hr_pay.py the
        (label, key, width) tuple is unpacked three times and a different
        member is dead each time — worth naming rather than silencing.
  RUF005 (10) `a + [b]` -> `[*a, b]`. Both forms build a new list/tuple;
        no operand is mutated at any of the ten sites.
  RUF046 (4) `int(round(x))` -> `round(x)`. Single-argument round() on a
        float already returns an int, so the cast was a no-op — including
        the tiyin conversion in uzpay/common.py, where the value and the
        type are both unchanged.

ruff: 44 -> 23.

+++ stabler/api/reports.py
@@ -2161,7 +2161,7 @@ def get_sales_detail_report(
         kg = flt(r["total_kg"])
         rate = flt(r["rate"])
         amt = flt(r["amount"])
-        boxes = int(round(kg / 20.0)) if kg > 0 else 0
+        boxes = round(kg / 20.0) if kg > 0 else 0
         box_kg = 20.0 if boxes > 0 else 0.0
 
         total_sales += amt
